libc6-<arch>: create merged-usr symlinks via preinst
There is a corner case on a merged-usr system that affects users
installing and then removing a libc6-<arch> package: the top-level
symlink gets removed, as the library is usually the last package to
install files there.

Add a preinst snippet that, if the system is merged-usr, creates the
top-level symlink used by the package. This way if it gets installed
and then removed, and then installed again, everything will work.
Note that this does not implement the full conversion, as that's the
responsibility of the usrmerge package (moving files, etc) - this
simply deals with the common case of the missing symlink.

Once all packages ship only files under /usr, hopefully in Trixie,
this can be removed.

Closes: #926699